Refactorings
Removed leftovers from tf 1.x based implemention
